fixes #488: refactor reftranscriptdata and altseqbuilder to not use closures or attrs
